requirements.txt
configparser==3.5.0
decorator==4.0.10
enum34==1.1.6
flake8==3.0.4
Flask==0.10.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mccabe==0.5.2
pbr==1.10.0
pluggy==0.4.0
psycopg2==2.6.2
py==1.4.31
pycodestyle==2.0.0
pyflakes==1.2.3
six==1.10.0
SQLAlchemy==1.0.14
sqlalchemy-migrate==0.10.0
sqlparse==0.2.1
Tempita==0.5.2
tox==2.4.1
virtualenv==15.0.3
Werkzeug==0.11.11
WTForms==2.1
